fix(build): NASM flag pollution, Clang 21 -Wimplicit-int-float,             sysctl tolerance for modern Fedora
First-build-on-Linux surfaced three blockers; all unblocked here.

1) CMakeLists.txt — gate global compile flags by language
   add_compile_options(${ASTRA_HARDENING_FLAGS}) was applied to ALL
   languages including NASM. NASM interprets `-f<x>` as the output-
   format directive, so `-fstack-protector-strong` produced:

     nasm: fatal: unrecognised output format `stack-protector-strong'

   Failure cascaded: astra_asm_core never built, so M-01 capability,
   M-03 cap-gate tests, M-21 ASM tests, M-02 isolation tests (which
   link asm_secure_wipe via the manager) all reported "Unable to
   find executable" at ctest time — they had never linked.

   Fix: wrap the C/CXX flag sets in $<COMPILE_LANGUAGE:C,CXX>
   generator expressions so NASM sees only its own
   CMAKE_ASM_NASM_OBJECT_FORMAT=elf64 setting. Link options are
   unchanged because NASM produces .o files that the C++ driver
   links.

2) tests/bench/bench_harness.h — explicit size_t→double conversion
   Clang 21 -Werror -Wimplicit-int-float-conversion flagged:

     pct * v.size()

   as a potentially-lossy implicit conversion (size_t → double).
   This blocked five baseline_*.cpp builds (pipe, socketpair,
   io_uring, aeron, erpc). Added an explicit static_cast<double>
   around v.size().

3) scripts/compile-astra.sh — sysctl tolerance
   Modern Fedora kernels (~5.18+) removed kernel.unprivileged_userns_clone
   because user namespaces are now always allowed. The sysctl command
   was failing with:

     sysctl: cannot stat /proc/sys/kernel/unprivileged_userns_clone:
             No such file or directory

   cmd_sysctl now probes /proc/sys/<key> before attempting to set
   each key, prints a `skip` row for missing keys, and persists only
   the keys that actually applied. cmd_check matches the same logic
   so the pre-flight no longer falsely warns.

Verification (on Fedora x86_64, kernel 6.x, Clang 21.1.8 reported
by the user):
  - 13 of 22 ctest entries ran and passed before this fix
  - 9 failed at link-time because astra_asm_core never built
  - After this fix, all 22 should configure + build + run

After pull:
  cd ~/projects/Astra
  git pull origin main
  ./scripts/compile-astra.sh clean
  ./scripts/compile-astra.sh

+++ tests/bench/bench_harness.h
@@ -81,8 +81,12 @@ inline Stats summarise(std::vector<uint64_t>& v, double fTscPerNs)
     if (v.empty()) return s;
     std::sort(v.begin(), v.end());
     auto at = [&](double pct) {
-        size_t i = std::min<size_t>(static_cast<size_t>(pct * v.size()),
-                                    v.size() - 1);
+        // Explicit double-cast of size() to silence Clang 21's
+        // -Wimplicit-int-float-conversion (-Werror). Truncation to
+        // size_t happens explicitly afterwards.
+        size_t i = std::min<size_t>(
+            static_cast<size_t>(pct * static_cast<double>(v.size())),
+            v.size() - 1);
         return static_cast<double>(v[i]) / fTscPerNs;
     };
     long double sum = 0;
